services/functions/hook/requirements.txt
gitpython==2.1.9
requests==2.18.4
